sys: add `openssl` feature
If `bundled`, will add the appropriate CMake flag to the build. If not,
attempts to link with the proper library that pkg-config found. Also
available is the `bundled_openssl` feature, to tell `openssl-sys` to
build from source and statically link.

One caveat is that I haven't quite figured out how to separate bundled
openssl forcing pthread support (at least from the `openssl-sys` layer),
so for now `bundled_openssl` implies `threading`.