Add optional layer for embedding into tokio
Adds `AsyncBase` through the `tokio` feature flag, which allows running
within a tokio-based async context, cooperatively with other async
tasks.

Utilizes a forked version of libevent which splits out support for
turning the event_base manually, and injecting the I/O fd into tokio via
`AsyncFd`. It's quite a hack, but there isn't really another way to do
this, and at least the internal mechanics are kept in-tact as much as
possible.